requirements.txt
Django==2.2.23
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
MarkupPy==1.14
MarkupSafe==2.0.1
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
